@@ -15,7 +15,7 @@
   },
   "homepage": "https://github.com/LWJGL/lwjgl3-www#readme",
   "dependencies": {
-    "aws-sdk": "2.613.0",
+    "aws-sdk": "2.614.0",
     "body-parser": "1.19.0",
     "chalk": "3.0.0",
     "compression": "1.7.4",
@@ -38,11 +38,9 @@
     "@emotion/babel-plugin": "11.0.0-next.11",
     "@emotion/css": "11.0.0-next.11",
     "@fullhuman/postcss-purgecss": "2.0.6",
-    "@prettier/plugin-pug": "1.1.5",
-    "@reach/router": "1.3.1",
+    "@prettier/plugin-pug": "1.1.6",
     "@types/jszip": "3.1.7",
     "@types/lodash-es": "4.17.3",
-    "@types/reach__router": "1.2.6",
     "@types/react-dom": "16.9.5",
     "@types/scheduler": "0.16.1",
     "@typescript-eslint/eslint-plugin": "2.19.0",
@@ -74,19 +72,21 @@
     "postcss-preset-env": "6.7.0",
     "prettier": "1.19.1",
     "puppeteer": "2.1.1",
-    "react": "experimental",
-    "react-dom": "experimental",
+    "react": "0.0.0-experimental-241c4467e",
+    "react-dom": "0.0.0-experimental-241c4467e",
     "react-fast-compare": "3.0.1",
     "react-refresh": "0.7.2",
+    "react-router": "6.0.0-alpha.1",
+    "react-router-dom": "6.0.0-alpha.1",
     "react-svg-loader": "3.0.3",
     "sass-loader": "8.0.2",
     "scroll-into-view-if-needed": "2.2.24",
     "style-loader": "1.1.3",
     "terser": "4.6.3",
     "three": "0.113.2",
-    "typescript": "3.7.5",
+    "typescript": "3.8.1-rc",
     "typescript-styled-plugin": "0.15.0",
-    "webpack": "5.0.0-beta.12",
+    "webpack": "5.0.0-beta.13",
     "webpack-dev-middleware": "3.7.2",
     "webpack-hot-middleware": "2.25.0",
     "whatwg-fetch": "3.0.0"
