@@ -1,6 +1,6 @@
 import React from 'react';
 import { css } from '@emotion/css';
-import { Link } from '@reach/router';
+import { Link } from 'react-router-dom';
 import { Icon, Folder as FolderIcon } from '~/components/icons';
 
 export const FolderTH = css`
