@@ -1,27 +1,26 @@
 import React, { useEffect } from 'react';
+import { useParams } from 'react-router-dom';
 import { PageView } from '~/components/routes/PageView';
-import { RouteComponentProps, WindowLocation } from '@reach/router';
 import { Browser } from './components/Browser';
 import { PathResource } from './PathResource';
 
-interface BrowseRouteProps {
-  '*': string;
-}
-
-const BrowseRoute = (props: RouteComponentProps<BrowseRouteProps>) => {
-  const path = props['*'] !== undefined ? props['*'] : '';
+const BrowseRoute: React.FC<{ children?: never }> = () => {
+  const params = useParams();
+  const path = params['*'] !== undefined ? params['*'] : '';
 
   useEffect(() => {
     PathResource.load(path);
   }, [path]);
 
+  useEffect(() => {
+    document.body.style.background = 'gray';
+    return () => {
+      document.body.style.background = '';
+    };
+  }, []);
+
   return (
-    <PageView location={props.location as WindowLocation} title="Browse" description="Browse LWJGL files">
-      <style
-        dangerouslySetInnerHTML={{
-          __html: `body { background: gray }`,
-        }}
-      />
+    <PageView title="Browse" description="Browse LWJGL files">
       <section className="container px-0 bg-white shadow" style={{ marginTop: '-1em' }}>
         <Browser path={path} />
       </section>
